+++ libs/ui/src/lib/activities-table/activities-table.component.stories.ts
@@ -39,7 +39,6 @@ const activities: Activity[] = [
     updatedAt: new Date('2025-05-31T18:43:01.840Z'),
     userId: '081aa387-487d-4438-83a4-3060eb2a016e',
     account: {
-      balance: 150.2,
       comment: null,
       createdAt: new Date('2025-05-31T13:00:13.940Z'),
       currency: 'USD',
@@ -105,7 +104,6 @@ const activities: Activity[] = [
     updatedAt: new Date('2025-05-31T18:46:14.175Z'),
     userId: '081aa387-487d-4438-83a4-3060eb2a016e',
     account: {
-      balance: 150.2,
       comment: null,
       createdAt: new Date('2025-05-31T13:00:13.940Z'),
       currency: 'USD',
@@ -171,7 +169,6 @@ const activities: Activity[] = [
     updatedAt: new Date('2025-05-31T18:49:54.064Z'),
     userId: '081aa387-487d-4438-83a4-3060eb2a016e',
     account: {
-      balance: 150.2,
       comment: null,
       createdAt: new Date('2025-05-31T13:00:13.940Z'),
       currency: 'USD',
@@ -237,7 +234,6 @@ const activities: Activity[] = [
     updatedAt: new Date('2025-05-31T18:48:48.209Z'),
     userId: '081aa387-487d-4438-83a4-3060eb2a016e',
     account: {
-      balance: 150.2,
       comment: null,
       createdAt: new Date('2025-05-31T13:00:13.940Z'),
       currency: 'USD',
@@ -303,7 +299,6 @@ const activities: Activity[] = [
     updatedAt: new Date('2025-05-31T18:46:44.616Z'),
     userId: '081aa387-487d-4438-83a4-3060eb2a016e',
     account: {
-      balance: 150.2,
       comment: null,
       createdAt: new Date('2025-05-31T13:00:13.940Z'),
       currency: 'USD',